backup_feature/backup_function/requirements.txt
# Google Cloud SDK libraries
google-cloud-storage==2.16.0
google-cloud-sqlconnector==1.6.0

# MySQL connector
mysql-connector-python==8.3.0

# Data handling
pandas==2.2.1
fastavro==1.9.4